# src/main/frontend/public/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hamstrack REST API
  version: 0.2.0
  description: |
    The HTTP API behind the Hamstrack web app — everything the UI does goes
    through it, and it is available to users for scripting and integrations.

    **Authentication** — call `POST /api/auth/login` and send the returned
    access token as `Authorization: Bearer <token>` (use the **Authorize**
    button here). Tokens expire after 24 h; refresh via the rotating
    `refresh_token` cookie with `POST /api/auth/refresh`.

    **Access model** — a resource you cannot see returns `404`, whether it
    does not exist or you are not a member of its workspace. Lists are
    currently unpaginated (beta). The API is unversioned while in beta;
    breaking changes are announced in release notes.

    **Rate limits** — sensitive auth endpoints share a per-IP budget
    (15 requests/min by default), and repeated failed logins for one account
    trigger an exponential backoff. Both return `429` with a `Retry-After`
    header (seconds).
  contact:
    name: Hamstrack
    url: https://github.com/Zherikhov/easyTask
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://hamstrack.com/api
    description: Cloud
  - url: /api
    description: This instance (Cloud or self-hosted DC)
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register an account
      description: |
        Sends a verification email. `termsAccepted: true` is required when the
        instance has terms acceptance enabled (see `GET /meta`). Returns `403`
        when the operator disabled public signup.
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Registered — verification email sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Email is already registered
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email with the one-time token
      description: Exchanges the emailed token for a session (acts as a login).
      operationId: verifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyEmailRequest' }
      responses:
        '200': { $ref: '#/components/responses/AuthOk' }
        '400':
          description: Token is invalid, expired or already used
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/resend-verification:
    post:
      tags: [Auth]
      summary: Re-send the verification email
      description: Always returns `200` — no account enumeration.
      operationId: resendVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      description: |
        Returns an access token and sets the `refresh_token` cookie
        (`HttpOnly`, `SameSite=Strict`, path `/api/auth`). Unverified accounts
        get `403`.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/AuthOk' }
        '401':
          description: Invalid email or password
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '403':
          description: Email address is not verified
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the refresh token
      description: |
        Authenticated by the `refresh_token` cookie. Returns a fresh access
        token and a new cookie; the previous refresh token is invalidated.
      operationId: refresh
      security:
        - cookieAuth: []
      responses:
        '200': { $ref: '#/components/responses/AuthOk' }
        '400':
          description: Missing, invalid or expired refresh token
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out
      description: Revokes the refresh token and clears the cookie.
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        '204': { description: Logged out }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset link
      description: Always returns `200` — no account enumeration.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password with the emailed token
      description: Revokes all active sessions on success.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8, maxLength: 100 }
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '400':
          description: Token is invalid, expired or already used
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      operationId: me
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /api/auth/login` (24 h lifetime).
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating refresh token cookie set on login (30-day lifetime, path `/api/auth`).
...
  responses:
    AuthOk:
      description: Authenticated — access token in body, refresh cookie set
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthResponse' }
    Unauthorized:
      description: Missing, expired or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Forbidden:
      description: Authenticated but not allowed (insufficient role)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    ValidationError:
      description: Failed validation — see the `errors` map
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblemDetail' }
...
    TooManyRequests:
      description: Rate limited — retry after the number of seconds in the `Retry-After` header
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
    ProblemDetail:
      type: object
      description: RFC 9457 problem details
      properties:
        type: { type: string, example: 'about:blank' }
        title: { type: string, example: Conflict }
        status: { type: integer, example: 409 }
        detail: { type: string, example: Issue was modified by someone else — refresh and retry }
    ValidationProblemDetail:
      allOf:
        - $ref: '#/components/schemas/ProblemDetail'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties: { type: string }
              example: { email: must be a well-formed email address }
    MessageResponse:
      type: object
      properties:
        message: { type: string }
...
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 100 }
        displayName: { type: string, minLength: 2, maxLength: 100 }
        termsAccepted:
          type: boolean
          description: Required to be `true` when the instance enforces terms acceptance
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, description: One-time token from the verification email }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string, description: JWT for the Authorization header }
        expiresIn: { type: integer, format: int64, example: 86400, description: 'Access token lifetime, seconds' }
        userId: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
    MeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
...